% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.check_igplotdat}
\alias{.check_igplotdat}
\title{validate igplotdat object}
\usage{
.check_igplotdat(x)
}
\arguments{
\item{x}{an arbitrary R object}
}
\value{
checks whether it is a valid \code{igplotdat} object and if not throws exceptions
}
\description{
validate igplotdat object
}
