% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.create_igplotdat}
\alias{.create_igplotdat}
\title{constructor of igplotdat object}
\usage{
.create_igplotdat(v = data.table(), e = data.table(), l = logical())
}
\arguments{
\item{v}{vertex data (\code{data.table})}

\item{e}{edge data (\code{data.table})}

\item{l}{boolean; \code{TRUE} if layout information are included and zero otherwise}
}
\value{
a \code{igplotdat} object
}
\description{
constructor of igplotdat object
}
