% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_attr_extr.R
\name{.igraph_attr_extr}
\alias{.igraph_attr_extr}
\title{Extract attributes from graph and combine with layout}
\usage{
.igraph_attr_extr(
  g,
  v_attr = NULL,
  e_attr = NULL,
  what = c("both", "vertex", "edge")
)
}
\arguments{
\item{g}{an \code{igraph} object}

\item{v_attr}{the vertex attributes to extract}

\item{e_attr}{the edge attributes to extract}

\item{what}{what to return}
}
\value{
returns at most two \code{data.tables} in a list: one table has attributes of the nodes and the other attributes of edges
}
\description{
Extract attributes from graph and combine with layout
}
