% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.transform_procrustes}
\alias{.transform_procrustes}
\title{Procrustes Transform of Rectangular matrix}
\usage{
.transform_procrustes(X, target, translation = TRUE, scale = TRUE)
}
\arguments{
\item{X}{dense matrix to transform}

\item{target}{dense matrix that is the target of the transform}

\item{translation}{boolean; whether positions should be translated}

\item{scale}{boolean; whether positions should be scaled}
}
\value{
returns a list of four elements:
\enumerate{
\item the transformed matrix \code{Xstar}
\item the rotation matrix \code{R}
\item the translation vector \code{t_vec}
\item the scaling factor \code{sigma}
}
The transformed matrix \code{Xstar} can be recreated by the equation \code{sigma * X * R} if \code{translation} is \code{FALSE}. If \code{translation} is \code{TRUE}, adding \code{t_vec} to each row of the matrix \code{sigma * X * R} completes the transformation.
}
\description{
Procrustes Transform of Rectangular matrix
}
