% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igplot.R, R/zzz.R
\docType{package}
\name{igplot}
\alias{igplot}
\title{Function to plot network}
\usage{
igplot(
  g,
  layout = NULL,
  directed = FALSE,
  bg_col = "white",
  v_frame = "white",
  v_fill = "blue",
  v_lwd = 0.1,
  v_pch = 21,
  v_cex = 0.6,
  e_color = "grey",
  e_lwd = 0.3,
  e_length = 0.25,
  e_angle = 30,
  outfile = NULL,
  width = 7,
  height = 7,
  return_data = FALSE,
  plot = TRUE
)
}
\arguments{
\item{g}{an \code{igraph} object, a \code{matrix} or \code{data.table} object with two columns containing the edgelist, or a \code{igplotdat} object}

\item{layout}{either a string that is equal to the \code{igraph} function that creates the layout (e.g., "layout_with_fr") or a two-dimensional matrix of the coordinates of the vertices}

\item{directed}{boolean; if \code{g} is not an \code{igraph} object, indicates whether the edgelist is directed or not}

\item{bg_col}{background color}

\item{v_frame}{vertex frame color}

\item{v_fill}{vertex fill color}

\item{v_lwd}{vertex frame thickness}

\item{v_pch}{vertex pch}

\item{v_cex}{vertex size}

\item{e_color}{edge color}

\item{e_lwd}{edge linewidth}

\item{e_length}{length of the arrow when \code{directed} is \code{TRUE}}

\item{e_angle}{angle of the arrow when \code{directed} is \code{TRUE}}

\item{outfile}{path to file}

\item{width}{width of the pdf file (in inches)}

\item{height}{height of the pdf file (in inches)}

\item{return_data}{whether to return the data}

\item{plot}{whether to create the plot}
}
