% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_plot_prep.R
\name{igraph_plot_prep}
\alias{igraph_plot_prep}
\title{Prepare extracted attributes from \code{igraph} object for plotting}
\usage{
igraph_plot_prep(
  ve_list,
  layout,
  v_frame = "white",
  v_fill = "blue",
  v_lwd = 0.1,
  v_pch = 21,
  v_cex = 0.6,
  e_color = "grey",
  e_lwd = 0.3
)
}
\arguments{
\item{ve_list}{a \code{list} of two elements: a \code{data.table} of vertex attributes and a \code{data.table} of an edge-list attributes}

\item{layout}{a two-dimensional \code{array} of node positions}

\item{v_pch}{vertex shapes}

\item{v_cex}{vertex sizes}

\item{e_color}{edge colors}

\item{e_lwd}{line width of edges}

\item{v_color}{vertex colors}

\item{vf_cex_add}{how much to add (in percentages) to the vertex sizes when plotting frames}

\item{vf_color}{color of frames}

\item{vf_lwd}{line width of frames}
}
\description{
Prepare extracted attributes from \code{igraph} object for plotting
}
