% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igplotdat.R
\name{is.igplotdat}
\alias{is.igplotdat}
\title{check method}
\usage{
is.igplotdat(x)
}
\arguments{
\item{x}{arbitrary R object}
}
\value{
returns \code{TRUE} if \code{x} is a \code{igplotdat} object and zero otherwise.
}
\description{
check method
}
