% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{merge_layout}
\alias{merge_layout}
\title{helper function, merges vertex data to edge data}
\usage{
merge_layout(v, e)
}
\arguments{
\item{v}{vertex set}

\item{e}{edge set}
}
\value{
merges the coordinates in v into e
}
\description{
helper function, merges vertex data to edge data
}
