% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igplotdat.R
\name{print.igplotdat}
\alias{print.igplotdat}
\title{print method for \code{igplotdat} object}
\usage{
\method{print}{igplotdat}(x, ...)
}
\arguments{
\item{x}{an \code{igplotdat} object}

\item{...}{further arguments passed to print}
}
\description{
print method for \code{igplotdat} object
}
