% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procrustes.R
\name{procrustes}
\alias{procrustes}
\title{Procrustes-transform layout}
\usage{
procrustes(x, y)
}
\arguments{
\item{x}{layout to be transformed. Either a \code{data.table} object or a \code{matrix}}

\item{y}{target layout.Either a \code{data.table} object or a \code{matrix}}
}
\value{
transforms \code{y} to be as close as possible without changing the shape of the graph.
}
\description{
Procrustes-transform layout
}
