% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_plot_df.R
\name{subset_plot_df}
\alias{subset_plot_df}
\title{Extract subset from plot-ready list}
\usage{
subset_plot_df(x, sub_var, sub_val, keep_edges = c("all", "within", "none"))
}
\arguments{
\item{x}{list of data.tables to subset (generated by igraph_plot_prep)}

\item{sub_var}{variable to subset on; either a logical vector or a character string}

\item{sub_val}{value to subset on}

\item{keep_edges}{what edges to keep}
}
\description{
Extract subset from plot-ready list
}
\details{
if \code{sub_var} is a logical vector, it has to be of the same length as the number of rows in \code{x$vertices}. If \code{sub_var} is a string, it has to be equal a column in \code{x$vertices}.
}
