% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igplotdat.R
\name{.create_igplotdat}
\alias{.create_igplotdat}
\title{constructor of igplotdat object}
\usage{
.create_igplotdat(v = data.table(), e = data.table(), l = logical())
}
\description{
constructor of igplotdat object
}
