% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igplot_fun.R, R/zzz.R
\docType{package}
\name{igplot}
\alias{igplot}
\title{Function to plot network}
\usage{
igplot(
  g,
  layout = NULL,
  directed = FALSE,
  bg_col = "white",
  plot_opts = list(),
  outfile = NULL,
  width = 7,
  height = 7,
  return_data = FALSE,
  plot = TRUE,
  corp = FALSE
)
}
\arguments{
\item{g}{an \code{igraph} object, a \code{matrix} or \code{data.table} object with two columns containing the edgelist, or a \code{igplotdat} object}

\item{layout}{either a string that is equal to the \code{igraph} function that creates the layout (e.g., "layout_with_fr") or a two-dimensional matrix of the coordinates of the vertices}

\item{directed}{boolean; if \code{g} is not an \code{igraph} object, indicates whether the edgelist is directed or not}

\item{bg_col}{background color}

\item{plot_opts}{a named list containing plotting options (see details)}

\item{outfile}{path to file}

\item{width}{width of the pdf file (in inches)}

\item{height}{height of the pdf file (in inches)}

\item{return_data}{whether to return the data}

\item{plot}{whether to create the plot}

\item{corp}{boolean; if \code{TRUE} corps the image to the positions of nodes}
}
\description{
This function uses the base::plot function to plot igraph objects.
}
\details{
Currently, the following options are supported for the \code{plot_opts} list:
\enumerate{
\item \code{v_frame}, vertex frame color
\item \code{v_lwd}, vertex frame thickness
\item \code{v_fill}, vertex fill color
\item \code{v_cex}, vertex size
\item \code{v_pch}, vertex pch
\item \code{e_col}, edge color
\item \code{e_lwd}, edge linewidth
\item \code{e_length}, length of the arrow when \code{directed} is \code{TRUE}
\item \code{e_angle}, angle of the arrow when \code{directed} is \code{TRUE}
}
}
